+++ src/filters/inspect.c
@@ -500,7 +500,7 @@ void gf_inspect_dump_obu(FILE *dump, AV1State *av1, char *obu, u64 obu_length, O
 	fprintf(dump, "   <OBU size=\""LLU"\" type=\"%s\" header_size=\"%d\" has_size_field=\"%d\" has_ext=\"%d\" temporalID=\"%d\" spatialID=\"%d\" ",  obu_size, av1_get_obu_name(obu_type), hdr_size, av1->obu_has_size_field, av1->obu_extension_flag, av1->temporal_id , av1->spatial_id);
 
 	if (dump_crc && (obu_size<0xFFFFFFFFUL) )
-		fprintf(dump, "crc=\"%u\" ", gf_crc_32(obu, obu_length) );
+		fprintf(dump, "crc=\"%u\" ", gf_crc_32(obu, (u32) obu_length) );
 
 	switch (obu_type) {
 	case OBU_SEQUENCE_HEADER:
